Photorealistic PCB rendering on back side
A UX breaking change is introduced to the render.js script,
where the back side canvas is rotated by 180 degree, and without any
horizontal flip.

Such change is necessary because the back side rendering of the PCB is
already flipped horizontally.

Here, the photorealistic rendering of the PCB is assumed to be exported
by the PCB software as 'basename-front.png', 'basename-back.png' at a
specific millimeter-per-pixel scale. The absolute coordinate of each
component, as defined in the centroid file (basename.xy), should match
the corresponding rendered element of the PCB.